requirements.txt
requests==2.31.0
python-dotenv==1.0.0
flask==3.0.2
geopy==2.4.1
certifi==2024.2.2
pydantic==2.6.3 